healthcare_voice/ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.27.0
websockets==12.0
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
openai==1.12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
twilio==8.10.0
